# WebP - Reduce the allocations in lossless encoding