… src/ImageSharp/Formats/Webp/Lossless/Vp8LHistogram.cs
… src/ImageSharp/Formats/Webp/Lossless/Vp8LHistogramSet.cs
… tests/ImageSharp.Benchmarks/Codecs/Webp/EncodeWebp.cs
… tests/ImageSharp.Tests/Formats/WebP/DominantCostRangeTests.cs
… tests/ImageSharp.Tests/Formats/WebP/Vp8LHistogramTests.cs