@JimBobSquarePants 2023-10-23
@antonfirsov I want to get this merged so it's not blocking #2569 

None of the changes here represent security enhancements so there's nothing to backport. Any future ones can be manually handled.